Fix verify_password test
The test case assumed that user.name is `user0`, which is no longer
the case now that we're adding additional tests.

Also, the test case assumed that `flask.g.current_user` is not set if
the password is invalid. This assumption does not hold true. It will
be set as long as the username is valid.