Disable FLASK_DEBUG when running tests
When the application is in debug mode the request context is not
popped on exceptions to enable debuggers to introspect the data. This
can result in `AssertionError: Popped wrong request context`, when an
exception is raised during a test case.

The alternative workaround would be to add
`PRESERVE_CONTEXT_ON_EXCEPTION = False` to `tests.config`.