Configure flake8 for black and flake8-bugbear
 - Use line length of 80
 - Disable E501, enable B950
 - Disable E203
 - Disable W503

The flake8 warning "line too long" (E501) is disabled in favor of
flake-bugbear "Line too long" (B950). This considers max-line-length
but only triggers when the value has been exceeded by more than 10%.

The flake8 warning "whitespace before ':'" (E203) is not PEP8
compliant. Black will treat : in slices as a binary operator with the
lowest priority, and leave an equal amount of space on either side,
except if a parameter is omitted.

The flake8 warning "line break before binary operator" (W503) is not
PEP8 compliant. Black will break a line before a binary operator when
splitting a block of code over multiple lines.

https://black.readthedocs.io/en/stable/the_black_code_style.html#slices
https://black.readthedocs.io/en/stable/the_black_code_style.html#line-breaks-binary-operators