Upgrade to marshmallow 3.0.0rc1
The following changes are most relevant to us:

 - Schemas are always strict
 - Schema.load and Schema.dump return only data, not errors
 - Deserialization fails on unknown keys

https://marshmallow.readthedocs.io/en/3.0/upgrading.html

This means that passing "id" on PUT and POST fails with status code
422. Our tests passed "id" frequently, because they used server-side
serialization and thus included "id".

We changed the expected behaviour for some tests, because encountering
"id" in the request data should result in 422 instead of silently
ignoring the field.