Allow poetry to detect virtualenv created by nox
Fix a bug where Poetry cannot detect the virtualenv created by nox. As a
result, Poetry uses its own virtualenv for the project, breaking the
isolation of the nox session.

This issue is caused by the fact that nox does not actually activate its
virtualenv when running a command. Instead, nox simply locates the
command to be run inside the virtualenv.

Pass the `VIRTUAL_ENV` environment variable to Poetry to allow it to
detect the virtualenv.